fix(core): Fix up hashmap serialization, write key/value pairs in interleaved fashion for easy read/write into hashmap + fixed const_itnerator bug.
- Fixed an issues with const interator, made const_interator point to const Key& andconst Value& instead
- Now hashmap serialization only supports trivial/primtive types
- Uses interleaved key/value pairs

+++ Tests/EngineTests/CoreTests/AssetDBTests.cpp
@@ -133,10 +133,10 @@ namespace Razix {
         const u32 headerB  = static_cast<u32>(handleB & RAZIX_ASSET_HOTDATA_MASK);
         const u32 payloadB = static_cast<u32>((handleB & RAZIX_ASSET_PAYLOLAD_INDEX_MASK) >> RAZIX_ASSET_PAYLOAD_SHIFT_INDEX);
 
-        EXPECT_LT(headerA, RAZIX_MAX_ASSETS);
-        EXPECT_LT(payloadA, RAZIX_MAX_ASSETS);
-        EXPECT_LT(headerB, RAZIX_MAX_ASSETS);
-        EXPECT_LT(payloadB, RAZIX_MAX_ASSETS);
+        EXPECT_LT(headerA, static_cast<u32>(RAZIX_MAX_ASSETS));
+        EXPECT_LT(payloadA, static_cast<u32>(RAZIX_MAX_ASSETS));
+        EXPECT_LT(headerB, static_cast<u32>(RAZIX_MAX_ASSETS));
+        EXPECT_LT(payloadB, static_cast<u32>(RAZIX_MAX_ASSETS));
 
         EXPECT_NE(headerA, headerB);
         EXPECT_NE(payloadA, payloadB);
